+++ changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * New borrowing validation tests.
     * End borrowing validation tests.
     * Add inventory item validation tests.
+    * Delete inventory item validation tests.
 * Models factories
     * Users.
     * Inventory items.